added cast to srand
git-svn-id: svn://svn.code.sf.net/p/axtls/code/trunk@228 9a5d90b5-6617-0410-8a86-bb477d3ed2e3

+++ crypto/crypto_misc.c
@@ -124,7 +124,7 @@ EXP_FUNC void STDCALL RNG_initialize()
     /* start of with a stack to copy across */
     int i;
     memcpy(entropy_pool, &i, ENTROPY_POOL_SIZE);
-    srand(&i); 
+    srand((unsigned int)&i); 
 #endif
 }
